Create the separate Icon file.
