+++ frontend/src/components/problem/ProblemDisplay.tsx
@@ -27,9 +27,8 @@ import {
   GreenSmallButtonBlock,
   InvertedSmallButton,
   TextButton,
-  InlineLobbyIcon,
-  InlineErrorIcon,
 } from '../core/Button';
+import { InlineIcon, InlineErrorIcon } from '../core/Icon';
 import ToggleButton from '../core/ToggleButton';
 import PrimarySelect from '../core/Select';
 import {
@@ -167,7 +166,7 @@ const TextButtonLink = styled(TextButton)`
   text-align: left;
 `;
 
-const InlineProblemIcon = styled(InlineLobbyIcon)`
+const InlineProblemIcon = styled(InlineIcon)`
   margin: 10px;
   height: 1rem;
 `;
